+++ examples/03_noprefetch/src/Item.tsx
@@ -2,7 +2,7 @@ import React, { useLayoutEffect, useRef, useState } from 'react';
 
 import { useFetcher } from 'react-hooks-fetch';
 
-import { fetcher } from './App';
+import { fetcher, initialSuspendable } from './App';
 import DisplayData from './DisplayData';
 
 type Props = {
@@ -11,14 +11,14 @@ type Props = {
 
 const Item: React.FC<Props> = ({ initialId = '' }) => {
   const [id, setId] = useState('');
-  const result = useFetcher(fetcher);
+  const result = useFetcher(fetcher, initialSuspendable);
   const initialized = useRef(false);
   useLayoutEffect(() => {
     if (!initialized.current && initialId !== id) {
-      initialized.current = true;
       setId(initialId);
       result.refetch(initialId);
     }
+    initialized.current = true;
   }, [initialId, id, result]);
   return (
     <div>